Add div that displays results of each round
This commit adds a div container that displays the results of each
round played, after the the player clicks a button and chooses their
play. This addition allows the user to immediately receive feedback on
how the game is progressing instead of having to navigate to the
console to do so.